requirements.txt
asgiref==3.11.0
colorama==0.4.6
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
whitenoise==6.11.0
et_xmlfile==2.0.0
iniconfig==2.3.0
openpyxl==3.1.5
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.2
python-dotenv==1.2.1
sqlparse==0.5.4
tzdata==2025.2
